implemented calculate total monthly cost functionality, refactored Service model

+++ src/main/java/com/ninjaone/backendinterviewproject/service/domain/Service.java
@@ -32,7 +32,7 @@ public class Service {
     private LocalDateTime updatedDate;
 
     @Setter
-    private Integer cost;
+    private Long cost;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, updatable = false)
